discord_requirements.txt
# TeleSync Discord Downloader Dependencies
discord.py>=2.3.0
aiohttp>=3.8.0
aiofiles>=23.0.0
yt-dlp>=2023.1.0
asyncio
pathlib

